test(frontend): #835 대형 페이지 테스트를 시나리오 단위로 분리
