+++ frontend/src/pages/workspace/ui/WorkspaceSimulationPage.test.tsx
@@ -1,354 +1,17 @@
-import { fireEvent, render, screen, waitFor } from "@testing-library/react";
-import { beforeEach, describe, expect, it, vi } from "vitest";
-import { MemoryRouter, Route, Routes } from "react-router-dom";
-import { toast } from "sonner";
-
-import { WorkspaceSimulationPage } from "./WorkspaceSimulationPage";
+import { fireEvent, screen, waitFor } from "@testing-library/react";
+import { describe, expect, it } from "vitest";
 import {
-  simulationApi,
-  type SimulationFeedbackType,
-  type SimulationImprovementCandidate,
-} from "@/features/simulation";
-import { useListAllWorkspaceWorkflows } from "@/entities/workflow";
-
-const setCrumbs = vi.fn();
-
-vi.mock("react-router-dom", async () => {
-  const actual = await vi.importActual<typeof import("react-router-dom")>("react-router-dom");
-  return {
-    ...actual,
-    useOutletContext: () => ({
-      setCrumbs,
-      workspace: { id: 1, name: "CS Team" },
-    }),
-  };
-});
-
-vi.mock("@/entities/workflow", () => ({
-  useListAllWorkspaceWorkflows: vi.fn(),
-}));
-
-vi.mock("@/features/simulation", () => ({
-  simulationApi: {
-    listSessions: vi.fn(),
-    createSession: vi.fn(),
-    getSession: vi.fn(),
-    sendMessage: vi.fn(),
-    createFeedback: vi.fn(),
-    listFeedback: vi.fn(),
-    createImprovementCandidate: vi.fn(),
-    listImprovementCandidates: vi.fn(),
-    updateImprovementCandidateStatus: vi.fn(),
-    approveImprovementCandidate: vi.fn(),
-    rejectImprovementCandidate: vi.fn(),
-    listGoldenCases: vi.fn(),
-    createGoldenCase: vi.fn(),
-    replayGoldenCase: vi.fn(),
-    listGoldenCaseReplays: vi.fn(),
-  },
-}));
-
-vi.mock("sonner", () => ({
-  toast: {
-    success: vi.fn(),
-    error: vi.fn(),
-  },
-}));
-
-const mockedWorkflows = vi.mocked(useListAllWorkspaceWorkflows);
-const mockedSimulationApi = vi.mocked(simulationApi);
-
-const session = {
-  id: 10,
-  channel: "SIMULATION",
-  status: "OPEN",
-  metaJson: JSON.stringify({ customerName: "테스트 고객" }),
-  startedAt: "2026-06-04T10:30:00Z",
-};
-
-const otherSession = {
-  id: 20,
-  channel: "SIMULATION",
-  status: "OPEN",
-  metaJson: JSON.stringify({ customerName: "다른 고객" }),
-  startedAt: "2026-06-04T11:30:00Z",
-};
-
-const detail = {
+  detail,
+  feedbackWithType,
+  mockedSimulationApi,
+  mockedWorkflows,
+  openCandidateTab,
+  openFeedbackTab,
+  otherSession,
+  renderPage,
   session,
-  messages: [
-    {
-      id: 1,
-      seqNo: 1,
-      senderRole: "USER",
-      messageType: "TEXT",
-      content: "환불하고 싶어요",
-      createdAt: "2026-06-04T10:31:00Z",
-    },
-  ],
-  matchedWorkflow: {
-    sessionId: 10,
-    workspaceId: 1,
-    domainPackId: 11,
-    domainPackVersionId: 101,
-    executionId: 77,
-    intentName: "환불 문의",
-    intentCode: "refund_request",
-    workflowDefinitionId: 100,
-    workflowCode: "refund_workflow",
-    workflowName: "환불 처리",
-    currentState: "collect_order_no",
-    executionStatus: "ACTIVE",
-  },
-  slotValues: { orderNo: "A-100" },
-  slots: [],
-  feedback: {
-    items: [
-      {
-        id: 900,
-        workspaceId: 1,
-        sessionId: 10,
-        chatMessageId: 1,
-        feedbackType: "MISSING_SLOT_QUESTION",
-        description: "주문번호를 묻지 않았습니다.",
-        expectedBehavior: "주문번호를 먼저 요청합니다.",
-        severity: "HIGH",
-        status: "OPEN",
-        createdBy: 7,
-        createdAt: "2026-06-04T10:40:00Z",
-        updatedAt: "2026-06-04T10:40:00Z",
-      },
-    ],
-    messageFeedbackCounts: { "1": 1 },
-  },
-};
-
-const candidate = {
-  id: 1000,
-  workspaceId: 1,
-  domainPackVersionId: 101,
-  feedbackId: 900,
-  sessionId: 10,
-  chatMessageId: 1,
-  candidateType: "SLOT_QUESTION",
-  targetElementType: "SLOT",
-  targetElementId: null,
-  targetElementKey: null,
-  beforeSummary: "주문번호를 묻지 않았습니다.",
-  afterSummary: "주문번호를 먼저 요청합니다.",
-  evidenceSummary: "simulation feedback #900",
-  reviewSessionId: null,
-  reviewTaskId: null,
-  appliedDomainPackVersionId: null,
-  draftPatchJson: "{}",
-  decisionReason: null,
-  decidedBy: null,
-  decidedAt: null,
-  status: "DRAFT",
-  createdBy: 7,
-  createdAt: "2026-06-04T10:45:00Z",
-  updatedAt: "2026-06-04T10:45:00Z",
-} as const;
-
-const validDraftPatchJson = JSON.stringify({
-  schemaVersion: "simulation-candidate-draft-patch.v1",
-  operation: "UPDATE_DESCRIPTION",
-  targetElementType: "SLOT",
-  targetElementId: 55,
-  targetElementKey: "orderNo",
-  beforeSummary: "주문번호를 묻지 않았습니다.",
-  afterSummary: "주문번호를 먼저 요청합니다.",
-  evidenceSummary: "simulation feedback #900 (session #10, turn #1)",
-});
-
-const arrayDraftPatchJson = JSON.stringify({
-  changes: [
-    null,
-    { field: "빈 변경" },
-    {
-      beforeValue: { required: false },
-      afterValue: true,
-    },
-  ],
-});
-
-const goldenCase = {
-  id: 950,
-  workspaceId: 1,
-  sourceSessionId: 10,
-  sourceDomainPackVersionId: 101,
-  name: "환불 검증",
-  inputMessagesJson: '[{"content":"환불하고 싶어요"}]',
-  expectedJson:
-    '{"intentCode":"refund_request","workflowCode":"refund_workflow","currentState":"collect_order_no","actionType":"ASK_SLOT"}',
-  createdBy: 7,
-  createdAt: "2026-06-05T10:45:00Z",
-  updatedAt: "2026-06-05T10:45:00Z",
-  latestReplayResult: null,
-} as const;
-
-const replayResult = {
-  id: 990,
-  workspaceId: 1,
-  goldenCaseId: 950,
-  domainPackVersionId: 101,
-  replaySessionId: 960,
-  status: "PASS",
-  expectedJson: "{}",
-  actualJson: "{}",
-  failureSummary: null,
-  createdBy: 7,
-  createdAt: "2026-06-05T10:46:00Z",
-} as const;
-
-const failedReplayResult = {
-  ...replayResult,
-  id: 991,
-  status: "FAIL",
-  failureSummary: "currentState expected collect_order_no but was handoff",
-} as const;
-
-function candidateWithType(
-  id: number,
-  candidateType: SimulationImprovementCandidate["candidateType"],
-): SimulationImprovementCandidate {
-  return {
-    ...candidate,
-    id,
-    candidateType,
-  };
-}
-
-function feedbackWithType(id: number, feedbackType: SimulationFeedbackType) {
-  return {
-    ...detail.feedback.items[0],
-    id,
-    feedbackType,
-  };
-}
-
-const otherDetail = {
-  ...detail,
-  session: otherSession,
-  messages: [
-    {
-      id: 3,
-      seqNo: 1,
-      senderRole: "USER",
-      messageType: "TEXT",
-      content: "배송지를 바꾸고 싶어요",
-      createdAt: "2026-06-04T11:31:00Z",
-    },
-  ],
-  feedback: {
-    items: [],
-    messageFeedbackCounts: {},
-  },
-};
-
-function renderPage(path = "/workspaces/1/simulation", state?: unknown) {
-  render(
-    <MemoryRouter initialEntries={[state === undefined ? path : { pathname: path, state }]}>
-      <Routes>
-        <Route path="/workspaces/:workspaceId/simulation" element={<WorkspaceSimulationPage />} />
-        <Route path="/workspaces" element={<div data-testid="workspace-root" />} />
-      </Routes>
-    </MemoryRouter>,
-  );
-}
-
-async function openFeedbackTab() {
-  fireEvent.click(await screen.findByRole("tab", { name: "피드백" }));
-}
-
-async function openCandidateTab() {
-  fireEvent.click(await screen.findByRole("tab", { name: "개선 후보" }));
-}
-
-beforeEach(() => {
-  vi.clearAllMocks();
-  mockedWorkflows.mockReturnValue({
-    loading: false,
-    error: null,
-    entries: [
-      {
-        packId: 11,
-        packName: "CS Support",
-        versionId: 22,
-        workflowId: 100,
-        workflowCode: "refund.standard",
-        name: "환불 처리",
-        description: null,
-        intentDefinitionId: 30,
-      },
-    ],
-  });
-  mockedSimulationApi.listSessions.mockResolvedValue({
-    content: [session],
-    page: 0,
-    size: 20,
-    totalElements: 1,
-    totalPages: 1,
-  });
-  mockedSimulationApi.getSession.mockImplementation(async (_workspaceId, sessionId) =>
-    sessionId === otherSession.id ? otherDetail : detail,
-  );
-  mockedSimulationApi.createSession.mockResolvedValue(detail);
-  mockedSimulationApi.createFeedback.mockResolvedValue(detail);
-  mockedSimulationApi.listFeedback.mockResolvedValue({
-    content: detail.feedback.items,
-    page: 0,
-    size: 20,
-    totalElements: 1,
-    totalPages: 1,
-  });
-  mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-    content: [],
-    page: 0,
-    size: 20,
-    totalElements: 0,
-    totalPages: 0,
-  });
-  mockedSimulationApi.createImprovementCandidate.mockResolvedValue(candidate);
-  mockedSimulationApi.updateImprovementCandidateStatus.mockResolvedValue({
-    ...candidate,
-    status: "READY_FOR_REVIEW",
-  });
-  mockedSimulationApi.approveImprovementCandidate.mockResolvedValue({
-    ...candidate,
-    status: "APPLIED",
-    appliedDomainPackVersionId: 102,
-    decisionReason: "시뮬레이션 리뷰 승인",
-  });
-  mockedSimulationApi.rejectImprovementCandidate.mockResolvedValue({
-    ...candidate,
-    status: "REJECTED",
-    decisionReason: "근거가 부족합니다.",
-  });
-  mockedSimulationApi.listGoldenCases.mockResolvedValue({
-    content: [],
-    page: 0,
-    size: 20,
-    totalElements: 0,
-    totalPages: 0,
-  });
-  mockedSimulationApi.createGoldenCase.mockResolvedValue(goldenCase);
-  mockedSimulationApi.replayGoldenCase.mockResolvedValue(replayResult);
-  mockedSimulationApi.sendMessage.mockResolvedValue({
-    ...detail,
-    messages: [
-      ...detail.messages,
-      {
-        id: 2,
-        seqNo: 2,
-        senderRole: "ASSISTANT",
-        messageType: "TEXT",
-        content: "주문번호를 알려주세요.",
-        createdAt: "2026-06-04T10:32:00Z",
-      },
-    ],
-  });
-});
+  toast,
+} from "./WorkspaceSimulationPage.test-helper";
 
 describe("WorkspaceSimulationPage", () => {
   it("잘못된 workspaceId면 /workspaces로 리다이렉트한다", () => {
@@ -359,7 +22,9 @@ describe("WorkspaceSimulationPage", () => {
   it("세션 목록과 runtime 상태를 표시한다", async () => {
     renderPage();
 
-    expect(await screen.findByRole("heading", { name: "상담 시뮬레이션" })).toBeInTheDocument();
+    expect(
+      await screen.findByRole("heading", { name: "상담 시뮬레이션" }),
+    ).toBeInTheDocument();
     expect(await screen.findByText("테스트 고객")).toBeInTheDocument();
     expect(await screen.findByText("환불하고 싶어요")).toBeInTheDocument();
     expect(screen.getByText("환불 문의")).toBeInTheDocument();
@@ -369,12 +34,18 @@ describe("WorkspaceSimulationPage", () => {
   });
 
   it("대시보드 추천 query로 피드백과 개선 후보 상태 필터를 초기화한다", async () => {
-    renderPage("/workspaces/1/simulation?feedbackStatus=RESOLVED&candidateStatus=READY_FOR_REVIEW");
+    renderPage(
+      "/workspaces/1/simulation?feedbackStatus=RESOLVED&candidateStatus=READY_FOR_REVIEW",
+    );
 
     await openFeedbackTab();
-    expect(await screen.findByLabelText("피드백 상태 필터")).toHaveValue("RESOLVED");
+    expect(await screen.findByLabelText("피드백 상태 필터")).toHaveValue(
+      "RESOLVED",
+    );
     await openCandidateTab();
-    expect(screen.getByLabelText("개선 후보 상태 필터")).toHaveValue("READY_FOR_REVIEW");
+    expect(screen.getByLabelText("개선 후보 상태 필터")).toHaveValue(
+      "READY_FOR_REVIEW",
+    );
     await waitFor(() => {
       expect(mockedSimulationApi.listFeedback).toHaveBeenCalledWith(1, {
         status: "RESOLVED",
@@ -383,7 +54,9 @@ describe("WorkspaceSimulationPage", () => {
       });
     });
     await waitFor(() => {
-      expect(mockedSimulationApi.listImprovementCandidates).toHaveBeenCalledWith(1, {
+      expect(
+        mockedSimulationApi.listImprovementCandidates,
+      ).toHaveBeenCalledWith(1, {
         status: "READY_FOR_REVIEW",
         page: 0,
         size: 20,
@@ -392,11 +65,17 @@ describe("WorkspaceSimulationPage", () => {
   });
 
   it("query 검증 대상을 상단에 표시하고 시작 workflow 기본값으로 사용한다", async () => {
-    renderPage("/workspaces/1/simulation?packId=11&versionId=22&workflowId=100");
+    renderPage(
+      "/workspaces/1/simulation?packId=11&versionId=22&workflowId=100",
+    );
 
     expect(await screen.findByText("Verification Target")).toBeInTheDocument();
-    expect(screen.getByRole("heading", { name: "환불 처리" })).toBeInTheDocument();
-    expect(screen.getByText("CS Support · Version #22 · refund.standard")).toBeInTheDocument();
+    expect(
+      screen.getByRole("heading", { name: "환불 처리" }),
+    ).toBeInTheDocument();
+    expect(
+      screen.getByText("CS Support · Version #22 · refund.standard"),
+    ).toBeInTheDocument();
     await waitFor(() => {
       expect(screen.getByLabelText("시작 workflow 선택")).toHaveValue("100");
     });
@@ -421,14 +100,18 @@ describe("WorkspaceSimulationPage", () => {
     });
 
     expect(await screen.findByText("Verification Target")).toBeInTheDocument();
-    expect(screen.getByRole("heading", { name: "환불 처리" })).toBeInTheDocument();
+    expect(
+      screen.getByRole("heading", { name: "환불 처리" }),
+    ).toBeInTheDocument();
     await waitFor(() => {
       expect(screen.getByLabelText("시작 workflow 선택")).toHaveValue("100");
     });
   });
 
   it("query 검증 대상이 있어도 사용자가 자동 매칭을 선택하면 workflow를 고정하지 않는다", async () => {
-    renderPage("/workspaces/1/simulation?packId=11&versionId=22&workflowId=100");
+    renderPage(
+      "/workspaces/1/simulation?packId=11&versionId=22&workflowId=100",
+    );
 
     const workflowSelect = await screen.findByLabelText("시작 workflow 선택");
     await waitFor(() => {
@@ -464,9 +147,12 @@ describe("WorkspaceSimulationPage", () => {
   it("고객 메시지를 전송하고 응답을 화면에 반영한다", async () => {
     renderPage();
 
-    fireEvent.change(await screen.findByPlaceholderText("고객 역할 메시지 입력"), {
-      target: { value: "A-100 주문 환불이요" },
-    });
+    fireEvent.change(
+      await screen.findByPlaceholderText("고객 역할 메시지 입력"),
+      {
+        target: { value: "A-100 주문 환불이요" },
+      },
+    );
     fireEvent.click(screen.getByRole("button", { name: "전송" }));
 
     await waitFor(() => {
@@ -474,14 +160,19 @@ describe("WorkspaceSimulationPage", () => {
         content: "A-100 주문 환불이요",
       });
     });
-    expect(await screen.findByText("주문번호를 알려주세요.")).toBeInTheDocument();
+    expect(
+      await screen.findByText("주문번호를 알려주세요."),
+    ).toBeInTheDocument();
   });
 
   it("turn 단위 피드백을 작성하고 session 상세를 갱신한다", async () => {
     renderPage();
 
     fireEvent.click(await screen.findByLabelText("Turn 1 피드백 대상 선택"));
-    expect(screen.getByRole("tab", { name: "피드백" })).toHaveAttribute("aria-selected", "true");
+    expect(screen.getByRole("tab", { name: "피드백" })).toHaveAttribute(
+      "aria-selected",
+      "true",
+    );
     expect(await screen.findByLabelText("피드백 대상 선택")).toHaveValue("1");
     fireEvent.change(screen.getByLabelText("피드백 유형 선택"), {
       target: { value: "MISSING_SLOT_QUESTION" },
@@ -540,7 +231,9 @@ describe("WorkspaceSimulationPage", () => {
     await waitFor(() => {
       expect(screen.getByLabelText("피드백 대상 선택")).toHaveValue("session");
     });
-    expect(screen.getByLabelText("피드백 유형 선택")).toHaveValue("INTENT_MISMATCH");
+    expect(screen.getByLabelText("피드백 유형 선택")).toHaveValue(
+      "INTENT_MISMATCH",
+    );
     expect(screen.getByLabelText("피드백 심각도 선택")).toHaveValue("MEDIUM");
     expect(screen.getByLabelText("설명")).toHaveValue("");
     expect(screen.getByLabelText("기대 응답/행동")).toHaveValue("");
@@ -557,63 +250,96 @@ describe("WorkspaceSimulationPage", () => {
     fireEvent.change(screen.getByLabelText("기대 응답/행동"), {
       target: { value: "주문번호를 먼저 요청합니다." },
     });
-    mockedSimulationApi.listFeedback.mockRejectedValueOnce(new Error("refresh failed"));
+    mockedSimulationApi.listFeedback.mockRejectedValueOnce(
+      new Error("refresh failed"),
+    );
     fireEvent.click(screen.getByRole("button", { name: "피드백 저장" }));
 
     await waitFor(() => {
       expect(mockedSimulationApi.createFeedback).toHaveBeenCalled();
     });
-    expect(toast.success).toHaveBeenCalledWith("시뮬레이션 피드백을 남겼습니다.");
+    expect(toast.success).toHaveBeenCalledWith(
+      "시뮬레이션 피드백을 남겼습니다.",
+    );
     await waitFor(() => {
-      expect(screen.getByText("시뮬레이션 피드백 목록을 불러오지 못했습니다.")).toBeInTheDocument();
+      expect(
+        screen.getByText("시뮬레이션 피드백 목록을 불러오지 못했습니다."),
+      ).toBeInTheDocument();
     });
-    expect(toast.error).not.toHaveBeenCalledWith("시뮬레이션 피드백 목록을 불러오지 못했습니다.");
-    expect(toast.error).not.toHaveBeenCalledWith("시뮬레이션 피드백을 저장하지 못했습니다.");
+    expect(toast.error).not.toHaveBeenCalledWith(
+      "시뮬레이션 피드백 목록을 불러오지 못했습니다.",
+    );
+    expect(toast.error).not.toHaveBeenCalledWith(
+      "시뮬레이션 피드백을 저장하지 못했습니다.",
+    );
   });
 
   it("피드백 목록 로드 실패는 패널 내부 오류와 재시도를 제공한다", async () => {
-    mockedSimulationApi.listFeedback.mockRejectedValueOnce(new Error("load failed"));
+    mockedSimulationApi.listFeedback.mockRejectedValueOnce(
+      new Error("load failed"),
+    );
     renderPage();
 
     await openFeedbackTab();
     expect(
       await screen.findByText("시뮬레이션 피드백 목록을 불러오지 못했습니다."),
     ).toBeInTheDocument();
-    expect(toast.error).not.toHaveBeenCalledWith("시뮬레이션 피드백 목록을 불러오지 못했습니다.");
+    expect(toast.error).not.toHaveBeenCalledWith(
+      "시뮬레이션 피드백 목록을 불러오지 못했습니다.",
+    );
 
     fireEvent.click(screen.getByRole("button", { name: "다시 시도" }));
 
-    expect(await screen.findByText("주문번호를 묻지 않았습니다.")).toBeInTheDocument();
+    expect(
+      await screen.findByText("주문번호를 묻지 않았습니다."),
+    ).toBeInTheDocument();
   });
 
   it("검증 케이스 목록 로드 실패는 패널 내부 오류로 표시한다", async () => {
-    mockedSimulationApi.listGoldenCases.mockRejectedValueOnce(new Error("load failed"));
+    mockedSimulationApi.listGoldenCases.mockRejectedValueOnce(
+      new Error("load failed"),
+    );
     renderPage();
 
-    expect(await screen.findByText("검증 케이스 목록을 불러오지 못했습니다.")).toBeInTheDocument();
-    expect(toast.error).not.toHaveBeenCalledWith("검증 케이스 목록을 불러오지 못했습니다.");
+    expect(
+      await screen.findByText("검증 케이스 목록을 불러오지 못했습니다."),
+    ).toBeInTheDocument();
+    expect(toast.error).not.toHaveBeenCalledWith(
+      "검증 케이스 목록을 불러오지 못했습니다.",
+    );
 
     fireEvent.click(screen.getByRole("button", { name: "다시 시도" }));
 
-    expect(await screen.findByText("저장된 검증 케이스가 없습니다.")).toBeInTheDocument();
+    expect(
+      await screen.findByText("저장된 검증 케이스가 없습니다."),
+    ).toBeInTheDocument();
   });
 
   it("OPEN 피드백에서 개선 후보를 생성하고 목록을 새로고침한다", async () => {
     renderPage();
 
     await openFeedbackTab();
-    expect(await screen.findByLabelText("피드백 #900 개선 대상 선택")).toHaveValue("SLOT");
+    expect(
+      await screen.findByLabelText("피드백 #900 개선 대상 선택"),
+    ).toHaveValue("SLOT");
     expect(screen.getByText("세부 element 미선택")).toBeInTheDocument();
-    expect(screen.getByText("현재 화면은 세부 element 선택을 지원하지 않아 target type까지만 후보에 저장합니다.")).toBeInTheDocument();
+    expect(
+      screen.getByText(
+        "현재 화면은 세부 element 선택을 지원하지 않아 target type까지만 후보에 저장합니다.",
+      ),
+    ).toBeInTheDocument();
     fireEvent.click(await screen.findByRole("button", { name: "후보" }));
 
     await waitFor(() => {
-      expect(mockedSimulationApi.createImprovementCandidate).toHaveBeenCalledWith(
+      expect(
+        mockedSimulationApi.createImprovementCandidate,
+      ).toHaveBeenCalledWith(
         1,
         900,
         expect.objectContaining({
           targetElementType: "SLOT",
-          beforeSummary: "Slot 개선 후보 (환불 처리 workflow 맥락): 주문번호를 묻지 않았습니다.",
+          beforeSummary:
+            "Slot 개선 후보 (환불 처리 workflow 맥락): 주문번호를 묻지 않았습니다.",
           afterSummary: "주문번호를 먼저 요청합니다.",
         }),
       );
@@ -621,7 +347,9 @@ describe("WorkspaceSimulationPage", () => {
     expect(toast.success).toHaveBeenCalledWith("개선 후보를 생성했습니다.");
     await waitFor(() => {
       expect(mockedSimulationApi.listFeedback).toHaveBeenCalledTimes(2);
-      expect(mockedSimulationApi.listImprovementCandidates).toHaveBeenCalledTimes(2);
+      expect(
+        mockedSimulationApi.listImprovementCandidates,
+      ).toHaveBeenCalledTimes(2);
     });
   });
 
@@ -629,22 +357,30 @@ describe("WorkspaceSimulationPage", () => {
     renderPage();
 
     await openFeedbackTab();
-    fireEvent.change(await screen.findByLabelText("피드백 #900 개선 대상 선택"), {
-      target: { value: "WORKFLOW" },
-    });
+    fireEvent.change(
+      await screen.findByLabelText("피드백 #900 개선 대상 선택"),
+      {
+        target: { value: "WORKFLOW" },
+      },
+    );
     expect(screen.getByText("#100 · refund_workflow")).toBeInTheDocument();
-    expect(screen.getByText("환불 처리 workflow id/key를 후보에 함께 저장합니다.")).toBeInTheDocument();
+    expect(
+      screen.getByText("환불 처리 workflow id/key를 후보에 함께 저장합니다."),
+    ).toBeInTheDocument();
     fireEvent.click(screen.getByRole("button", { name: "후보" }));
 
     await waitFor(() => {
-      expect(mockedSimulationApi.createImprovementCandidate).toHaveBeenCalledWith(
+      expect(
+        mockedSimulationApi.createImprovementCandidate,
+      ).toHaveBeenCalledWith(
         1,
         900,
         expect.objectContaining({
           targetElementType: "WORKFLOW",
           targetElementId: 100,
           targetElementKey: "refund_workflow",
-          beforeSummary: "Workflow 개선 후보 (환불 처리 workflow 맥락): 주문번호를 묻지 않았습니다.",
+          beforeSummary:
+            "Workflow 개선 후보 (환불 처리 workflow 맥락): 주문번호를 묻지 않았습니다.",
           afterSummary: "주문번호를 먼저 요청합니다.",
         }),
       );
@@ -655,18 +391,24 @@ describe("WorkspaceSimulationPage", () => {
     renderPage();
 
     await openFeedbackTab();
-    fireEvent.change(await screen.findByLabelText("피드백 #900 개선 대상 선택"), {
-      target: { value: "POLICY" },
-    });
+    fireEvent.change(
+      await screen.findByLabelText("피드백 #900 개선 대상 선택"),
+      {
+        target: { value: "POLICY" },
+      },
+    );
     fireEvent.click(screen.getByRole("button", { name: "후보" }));
 
     await waitFor(() => {
-      expect(mockedSimulationApi.createImprovementCandidate).toHaveBeenCalledWith(
+      expect(
+        mockedSimulationApi.createImprovementCandidate,
+      ).toHaveBeenCalledWith(
         1,
         900,
         expect.objectContaining({
           targetElementType: "POLICY",
-          beforeSummary: "Policy 개선 후보 (환불 처리 workflow 맥락): 주문번호를 묻지 않았습니다.",
+          beforeSummary:
+            "Policy 개선 후보 (환불 처리 workflow 맥락): 주문번호를 묻지 않았습니다.",
         }),
       );
     });
@@ -692,14 +434,32 @@ describe("WorkspaceSimulationPage", () => {
 
     await openFeedbackTab();
 
-    expect(await screen.findByLabelText("피드백 #901 개선 대상 선택")).toHaveValue("INTENT");
-    expect(screen.getByLabelText("피드백 #902 개선 대상 선택")).toHaveValue("SLOT");
-    expect(screen.getByLabelText("피드백 #903 개선 대상 선택")).toHaveValue("POLICY");
-    expect(screen.getByLabelText("피드백 #904 개선 대상 선택")).toHaveValue("RISK_RULE");
-    expect(screen.getByLabelText("피드백 #905 개선 대상 선택")).toHaveValue("WORKFLOW");
-    expect(screen.getByLabelText("피드백 #906 개선 대상 선택")).toHaveValue("RESPONSE");
-    expect(screen.getByLabelText("피드백 #907 개선 대상 선택")).toHaveValue("UNKNOWN");
-    expect(screen.getByText("기타 피드백은 구체 대상이 확정되지 않은 제한 상태로 후보화됩니다.")).toBeInTheDocument();
+    expect(
+      await screen.findByLabelText("피드백 #901 개선 대상 선택"),
+    ).toHaveValue("INTENT");
+    expect(screen.getByLabelText("피드백 #902 개선 대상 선택")).toHaveValue(
+      "SLOT",
+    );
+    expect(screen.getByLabelText("피드백 #903 개선 대상 선택")).toHaveValue(
+      "POLICY",
+    );
+    expect(screen.getByLabelText("피드백 #904 개선 대상 선택")).toHaveValue(
+      "RISK_RULE",
+    );
+    expect(screen.getByLabelText("피드백 #905 개선 대상 선택")).toHaveValue(
+      "WORKFLOW",
+    );
+    expect(screen.getByLabelText("피드백 #906 개선 대상 선택")).toHaveValue(
+      "RESPONSE",
+    );
+    expect(screen.getByLabelText("피드백 #907 개선 대상 선택")).toHaveValue(
+      "UNKNOWN",
+    );
+    expect(
+      screen.getByText(
+        "기타 피드백은 구체 대상이 확정되지 않은 제한 상태로 후보화됩니다.",
+      ),
+    ).toBeInTheDocument();
   });
 
   it("workflow 맥락이 없으면 workflow target type만 후보 payload에 보존한다", async () => {
@@ -723,12 +483,16 @@ describe("WorkspaceSimulationPage", () => {
 
     await openFeedbackTab();
     expect(
-      await screen.findByText("workflow 맥락이 확인되지 않아 target type만 후보에 저장됩니다."),
+      await screen.findByText(
+        "workflow 맥락이 확인되지 않아 target type만 후보에 저장됩니다.",
+      ),
     ).toBeInTheDocument();
     fireEvent.click(screen.getByRole("button", { name: "후보" }));
 
     await waitFor(() => {
-      expect(mockedSimulationApi.createImprovementCandidate).toHaveBeenCalledWith(
+      expect(
+        mockedSimulationApi.createImprovementCandidate,
+      ).toHaveBeenCalledWith(
         1,
         905,
         expect.objectContaining({
@@ -757,7 +521,9 @@ describe("WorkspaceSimulationPage", () => {
     fireEvent.click(screen.getByRole("button", { name: "후보" }));
 
     await waitFor(() => {
-      expect(mockedSimulationApi.createImprovementCandidate).toHaveBeenCalledWith(
+      expect(
+        mockedSimulationApi.createImprovementCandidate,
+      ).toHaveBeenCalledWith(
         1,
         900,
         expect.objectContaining({
@@ -767,568 +533,16 @@ describe("WorkspaceSimulationPage", () => {
     });
     expect(toast.success).toHaveBeenCalledWith("개선 후보를 생성했습니다.");
     await waitFor(() => {
-      expect(screen.getByText("개선 후보 목록을 불러오지 못했습니다.")).toBeInTheDocument();
+      expect(
+        screen.getByText("개선 후보 목록을 불러오지 못했습니다."),
+      ).toBeInTheDocument();
     });
-    expect(toast.error).not.toHaveBeenCalledWith("개선 후보 목록을 불러오지 못했습니다.");
-    expect(toast.error).not.toHaveBeenCalledWith("개선 후보를 생성하지 못했습니다.");
-  });
-
-  it("개선 후보 상태를 변경하고 후보 목록을 새로고침한다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [candidate],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    fireEvent.click(await screen.findByRole("button", { name: "리뷰 요청" }));
-
-    await waitFor(() => {
-      expect(mockedSimulationApi.updateImprovementCandidateStatus).toHaveBeenCalledWith(1, 1000, {
-        status: "READY_FOR_REVIEW",
-      });
-    });
-    expect(toast.success).toHaveBeenCalledWith("개선 후보 상태를 변경했습니다.");
-    await waitFor(() => {
-      expect(mockedSimulationApi.listImprovementCandidates).toHaveBeenCalledTimes(2);
-    });
-    expect(screen.getByRole("tab", { name: "개선 후보" })).toHaveAttribute("aria-selected", "true");
-  });
-
-  it("개선 후보 상태 변경 후 목록 새로고침 실패는 변경 실패로 처리하지 않는다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [candidate],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    const requestButton = await screen.findByRole("button", {
-      name: "리뷰 요청",
-    });
-    mockedSimulationApi.listImprovementCandidates.mockRejectedValueOnce(
-      new Error("refresh failed"),
+    expect(toast.error).not.toHaveBeenCalledWith(
+      "개선 후보 목록을 불러오지 못했습니다.",
     );
-    fireEvent.click(requestButton);
-
-    await waitFor(() => {
-      expect(mockedSimulationApi.updateImprovementCandidateStatus).toHaveBeenCalledWith(1, 1000, {
-        status: "READY_FOR_REVIEW",
-      });
-    });
-    expect(toast.success).toHaveBeenCalledWith("개선 후보 상태를 변경했습니다.");
-    await waitFor(() => {
-      expect(screen.getByText("개선 후보 목록을 불러오지 못했습니다.")).toBeInTheDocument();
-    });
-    expect(toast.error).not.toHaveBeenCalledWith("개선 후보 목록을 불러오지 못했습니다.");
-    expect(toast.error).not.toHaveBeenCalledWith("개선 후보 상태를 변경하지 못했습니다.");
-  });
-
-  it("개선 후보 유형 라벨을 표시한다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [
-        candidateWithType(1001, "INTENT_DESCRIPTION_EXAMPLE"),
-        candidateWithType(1002, "POLICY_CONDITION"),
-        candidateWithType(1003, "RISK_RULE"),
-        candidateWithType(1004, "WORKFLOW_STATE_TRANSITION"),
-        candidateWithType(1005, "HANDOFF_CONDITION"),
-        candidateWithType(1006, "RESPONSE_COPY"),
-        candidateWithType(1007, "OTHER"),
-        candidateWithType(1008, "CUSTOM" as SimulationImprovementCandidate["candidateType"]),
-      ],
-      page: 0,
-      size: 20,
-      totalElements: 8,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    expect(screen.getByRole("tab", { name: "개선 후보" })).toHaveAttribute("aria-selected", "true");
-    expect(await screen.findByText("intent 설명/예시")).toBeInTheDocument();
-    expect(screen.getByText("policy 조건")).toBeInTheDocument();
-    expect(screen.getByText("risk rule")).toBeInTheDocument();
-    expect(screen.getByText("workflow 전이")).toBeInTheDocument();
-    expect(screen.getByText("handoff 조건")).toBeInTheDocument();
-    expect(screen.getByText("응답 문구")).toBeInTheDocument();
-    expect(screen.getAllByText("기타")).toHaveLength(1);
-    expect(screen.getByText("CUSTOM")).toBeInTheDocument();
-    expect(screen.getAllByText("초안").length).toBeGreaterThan(0);
-    expect(screen.getAllByText("변경 전").length).toBeGreaterThan(0);
-    expect(screen.getAllByText("근거").length).toBeGreaterThan(0);
-  });
-
-  it("개선 후보 draft patch 변경 상세를 표시하고 확인 전 승인을 막는다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [
-        {
-          ...candidate,
-          status: "READY_FOR_REVIEW",
-          reviewSessionId: 200,
-          reviewTaskId: 300,
-          draftPatchJson: validDraftPatchJson,
-        },
-      ],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    expect(await screen.findByText("Draft patch 검토")).toBeInTheDocument();
-    expect(screen.getByText("UPDATE_DESCRIPTION")).toBeInTheDocument();
-    expect(screen.getAllByText("SLOT · #55 · orderNo").length).toBeGreaterThan(0);
-    expect(screen.getByText("초안 변경 요약")).toBeInTheDocument();
-    expect(screen.getAllByText("주문번호를 묻지 않았습니다.").length).toBeGreaterThan(0);
-    expect(screen.getAllByText("주문번호를 먼저 요청합니다.").length).toBeGreaterThan(0);
-    expect(screen.getByText("simulation feedback #900 (session #10, turn #1)")).toBeInTheDocument();
-
-    const approveButton = screen.getByRole("button", { name: "승인" });
-    expect(approveButton).toBeDisabled();
-
-    fireEvent.click(screen.getByRole("button", { name: "변경 상세 확인" }));
-    expect(screen.getByRole("button", { name: "변경 확인 완료" })).toBeDisabled();
-    expect(approveButton).not.toBeDisabled();
-  });
-
-  it("draft patch changes 배열의 변경값과 기본 표시를 검토한다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [
-        {
-          ...candidate,
-          status: "READY_FOR_REVIEW",
-          reviewSessionId: 200,
-          reviewTaskId: 300,
-          draftPatchJson: arrayDraftPatchJson,
-        },
-      ],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    expect(await screen.findByText("작업 미지정")).toBeInTheDocument();
-    expect(screen.getAllByText("SLOT").length).toBeGreaterThan(0);
-    expect(screen.getByText("변경 3")).toBeInTheDocument();
-    expect(screen.getByText('{"required":false}')).toBeInTheDocument();
-    expect(screen.getByText("true")).toBeInTheDocument();
-    expect(screen.getAllByText("simulation feedback #900").length).toBeGreaterThan(0);
-  });
-
-  it("draft patch 정보가 없으면 승인할 수 없음을 표시한다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [
-        {
-          ...candidate,
-          status: "READY_FOR_REVIEW",
-          reviewSessionId: 200,
-          reviewTaskId: 300,
-          draftPatchJson: "{}",
-        },
-      ],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    expect(
-      await screen.findByText("draft patch 정보가 없습니다. 변경 전/후 필드를 확인할 수 없습니다."),
-    ).toBeInTheDocument();
-    expect(screen.queryByRole("button", { name: "변경 상세 확인" })).not.toBeInTheDocument();
-    expect(screen.getByRole("button", { name: "승인" })).toBeDisabled();
-  });
-
-  it("draft patch가 객체 형식이 아니면 승인할 수 없음을 표시한다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [
-        {
-          ...candidate,
-          status: "READY_FOR_REVIEW",
-          reviewSessionId: 200,
-          reviewTaskId: 300,
-          draftPatchJson: '"not-object"',
-        },
-      ],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    expect(
-      await screen.findByText("draft patch가 객체 형식이 아니거나 비어 있습니다."),
-    ).toBeInTheDocument();
-    expect(screen.getByText('"not-object"')).toBeInTheDocument();
-    expect(screen.getByRole("button", { name: "승인" })).toBeDisabled();
-  });
-
-  it("draft patch JSON이 파싱되지 않으면 원문과 오류 상태를 표시한다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [
-        {
-          ...candidate,
-          status: "READY_FOR_REVIEW",
-          reviewSessionId: 200,
-          reviewTaskId: 300,
-          draftPatchJson: "{invalid-json",
-        },
-      ],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    expect(await screen.findByText("draft patch JSON을 해석할 수 없습니다.")).toBeInTheDocument();
-    expect(screen.getByText("{invalid-json")).toBeInTheDocument();
-    expect(screen.getByRole("button", { name: "승인" })).toBeDisabled();
-  });
-
-  it("개선 후보 목록 로드 실패는 패널 내부 오류와 재시도를 제공한다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockRejectedValueOnce(new Error("load failed"));
-    renderPage();
-
-    await openCandidateTab();
-    expect(await screen.findByText("개선 후보 목록을 불러오지 못했습니다.")).toBeInTheDocument();
-    expect(toast.error).not.toHaveBeenCalledWith("개선 후보 목록을 불러오지 못했습니다.");
-
-    fireEvent.click(screen.getByRole("button", { name: "다시 시도" }));
-
-    expect(await screen.findByText("조건에 맞는 개선 후보가 없습니다.")).toBeInTheDocument();
-  });
-
-  it("개선 후보 생성 실패를 토스트로 알린다", async () => {
-    renderPage();
-
-    await screen.findByText("환불하고 싶어요");
-    await openFeedbackTab();
-    mockedSimulationApi.createImprovementCandidate.mockRejectedValueOnce(
-      new Error("create failed"),
+    expect(toast.error).not.toHaveBeenCalledWith(
+      "개선 후보를 생성하지 못했습니다.",
     );
-    fireEvent.click(screen.getByRole("button", { name: "후보" }));
-
-    await waitFor(() => {
-      expect(toast.error).toHaveBeenCalledWith("개선 후보를 생성하지 못했습니다.");
-    });
-  });
-
-  it("개선 후보 상태 변경 실패를 토스트로 알린다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [candidate],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    const requestButton = await screen.findByRole("button", {
-      name: "리뷰 요청",
-    });
-    mockedSimulationApi.updateImprovementCandidateStatus.mockRejectedValueOnce(
-      new Error("update failed"),
-    );
-    fireEvent.click(requestButton);
-
-    await waitFor(() => {
-      expect(toast.error).toHaveBeenCalledWith("개선 후보 상태를 변경하지 못했습니다.");
-    });
-  });
-
-  it("READY_FOR_REVIEW 개선 후보를 승인하고 목록을 새로고침한다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [
-        {
-          ...candidate,
-          status: "READY_FOR_REVIEW",
-          reviewSessionId: 200,
-          reviewTaskId: 300,
-          draftPatchJson: validDraftPatchJson,
-        },
-      ],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    fireEvent.click(await screen.findByRole("button", { name: "변경 상세 확인" }));
-    fireEvent.click(await screen.findByRole("button", { name: "승인" }));
-
-    await waitFor(() => {
-      expect(mockedSimulationApi.approveImprovementCandidate).toHaveBeenCalledWith(1, 1000, {
-        reason: "시뮬레이션 리뷰 승인",
-      });
-    });
-    expect(toast.success).toHaveBeenCalledWith(expect.stringContaining("적용 version #102"));
-    expect(await screen.findByRole("status")).toHaveTextContent("적용 version #102");
-    expect(screen.getByRole("status")).toHaveTextContent("review session #200");
-    expect(screen.getByRole("status")).toHaveTextContent("review task #300");
-    await waitFor(() => {
-      expect(mockedSimulationApi.listImprovementCandidates).toHaveBeenCalledTimes(2);
-      expect(mockedSimulationApi.listFeedback).toHaveBeenCalledTimes(2);
-    });
-  });
-
-  it("반려 사유가 없으면 READY_FOR_REVIEW 개선 후보를 반려하지 않는다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [
-        {
-          ...candidate,
-          status: "READY_FOR_REVIEW",
-          reviewSessionId: 200,
-          reviewTaskId: 300,
-        },
-      ],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    fireEvent.click(await screen.findByRole("button", { name: "반려" }));
-
-    expect(toast.error).toHaveBeenCalledWith("반려 사유를 입력하세요.");
-    expect(mockedSimulationApi.rejectImprovementCandidate).not.toHaveBeenCalled();
-  });
-
-  it("READY_FOR_REVIEW 개선 후보를 반려하고 사유를 전달한다", async () => {
-    mockedSimulationApi.listImprovementCandidates.mockResolvedValue({
-      content: [
-        {
-          ...candidate,
-          status: "READY_FOR_REVIEW",
-          reviewSessionId: 200,
-          reviewTaskId: 300,
-        },
-      ],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await openCandidateTab();
-    fireEvent.change(await screen.findByLabelText("개선 후보 반려 사유"), {
-      target: { value: "근거가 부족합니다." },
-    });
-    fireEvent.click(screen.getByRole("button", { name: "반려" }));
-
-    await waitFor(() => {
-      expect(mockedSimulationApi.rejectImprovementCandidate).toHaveBeenCalledWith(1, 1000, {
-        reason: "근거가 부족합니다.",
-      });
-    });
-    expect(toast.success).toHaveBeenCalledWith("개선 후보를 반려했습니다.");
-  });
-
-  it("현재 실행 결과와 구분한 기대 결과를 검증 케이스로 저장한다", async () => {
-    renderPage();
-
-    await screen.findByText("환불하고 싶어요");
-    expect(screen.getByText("현재 실행 결과")).toBeInTheDocument();
-    expect(screen.getByText("기대 결과")).toBeInTheDocument();
-    fireEvent.change(await screen.findByLabelText("검증 케이스 이름"), {
-      target: { value: "환불 주문번호 검증" },
-    });
-    fireEvent.change(screen.getByLabelText("기대 intent"), {
-      target: { value: "refund_order_number_required" },
-    });
-    fireEvent.change(screen.getByLabelText("기대 workflow"), {
-      target: { value: "refund_required_slot_workflow" },
-    });
-    fireEvent.change(screen.getByLabelText("기대 state"), {
-      target: { value: "ask_order_no" },
-    });
-    fireEvent.change(await screen.findByLabelText("기대 action"), {
-      target: { value: "ASK_SLOT" },
-    });
-    fireEvent.change(screen.getByLabelText("필수 slot JSON"), {
-      target: { value: '{"orderNo":"B-200"}' },
-    });
-    fireEvent.click(screen.getByRole("button", { name: "등록" }));
-
-    await waitFor(() => {
-      expect(mockedSimulationApi.createGoldenCase).toHaveBeenCalledWith(1, 10, {
-        name: "환불 주문번호 검증",
-        expectedIntentCode: "refund_order_number_required",
-        expectedWorkflowCode: "refund_required_slot_workflow",
-        expectedCurrentState: "ask_order_no",
-        expectedActionType: "ASK_SLOT",
-        expectedSlotValues: { orderNo: "B-200" },
-      });
-    });
-    expect(toast.success).toHaveBeenCalledWith("검증 케이스를 저장했습니다.");
-  });
-
-  it("기대 intent, workflow, action 확인 전에는 검증 케이스를 저장하지 않는다", async () => {
-    renderPage();
-
-    await screen.findByText("환불하고 싶어요");
-    await waitFor(() => {
-      expect(screen.getByLabelText("기대 intent")).toHaveValue("refund_request");
-    });
-    fireEvent.change(screen.getByLabelText("기대 intent"), {
-      target: { value: "" },
-    });
-    fireEvent.change(screen.getByLabelText("기대 action"), {
-      target: { value: "ASK_SLOT" },
-    });
-    fireEvent.click(screen.getByRole("button", { name: "등록" }));
-
-    expect(toast.error).toHaveBeenCalledWith("기대 intent, workflow, action을 확인하세요.");
-    expect(mockedSimulationApi.createGoldenCase).not.toHaveBeenCalled();
-  });
-
-  it("필수 slot JSON이 객체가 아니면 검증 케이스를 저장하지 않는다", async () => {
-    renderPage();
-
-    await screen.findByText("환불하고 싶어요");
-    await waitFor(() => {
-      expect(screen.getByLabelText("기대 intent")).toHaveValue("refund_request");
-    });
-    fireEvent.change(screen.getByLabelText("기대 action"), {
-      target: { value: "ASK_SLOT" },
-    });
-    fireEvent.change(screen.getByLabelText("필수 slot JSON"), {
-      target: { value: '["orderNo"]' },
-    });
-    fireEvent.click(screen.getByRole("button", { name: "등록" }));
-
-    expect(toast.error).toHaveBeenCalledWith("필수 slot은 JSON 객체로 입력하세요.");
-    expect(mockedSimulationApi.createGoldenCase).not.toHaveBeenCalled();
-  });
-
-  it("저장된 검증 케이스를 선택 version으로 replay한다", async () => {
-    mockedSimulationApi.listGoldenCases.mockResolvedValue({
-      content: [goldenCase],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await screen.findByText("환불하고 싶어요");
-    await waitFor(() => {
-      expect(screen.getByLabelText("Replay version")).toHaveValue("101");
-    });
-    fireEvent.click(await screen.findByRole("button", { name: "환불 검증 replay" }));
-
-    await waitFor(() => {
-      expect(mockedSimulationApi.replayGoldenCase).toHaveBeenCalledWith(1, 950, {
-        domainPackVersionId: 101,
-      });
-    });
-    expect(toast.success).toHaveBeenCalledWith("검증 케이스 replay가 통과했습니다.");
-  });
-
-  it("query target version을 replay version 기본값으로 사용한다", async () => {
-    mockedSimulationApi.listGoldenCases.mockResolvedValue({
-      content: [goldenCase],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage("/workspaces/1/simulation?packId=11&versionId=22&workflowId=100");
-
-    await screen.findByText("환불하고 싶어요");
-    await waitFor(() => {
-      expect(screen.getByLabelText("Replay version")).toHaveValue("22");
-    });
-    fireEvent.click(await screen.findByRole("button", { name: "환불 검증 replay" }));
-
-    await waitFor(() => {
-      expect(mockedSimulationApi.replayGoldenCase).toHaveBeenCalledWith(1, 950, {
-        domainPackVersionId: 22,
-      });
-    });
-  });
-
-  it("최근 replay 실패 요약을 검증 케이스 목록에 표시한다", async () => {
-    mockedSimulationApi.listGoldenCases.mockResolvedValue({
-      content: [
-        {
-          ...goldenCase,
-          latestReplayResult: failedReplayResult,
-        },
-      ],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    expect(await screen.findByText("FAIL")).toBeInTheDocument();
-    expect(
-      screen.getByText("currentState expected collect_order_no but was handoff"),
-    ).toBeInTheDocument();
-  });
-
-  it("Replay version이 비어 있으면 replay 요청을 보내지 않는다", async () => {
-    mockedSimulationApi.listGoldenCases.mockResolvedValue({
-      content: [goldenCase],
-      page: 0,
-      size: 20,
-      totalElements: 1,
-      totalPages: 1,
-    });
-    renderPage();
-
-    await screen.findByRole("button", { name: "환불 검증 replay" });
-    await waitFor(() => {
-      expect(screen.getByLabelText("Replay version")).toHaveValue("101");
-    });
-    fireEvent.change(screen.getByLabelText("Replay version"), {
-      target: { value: "" },
-    });
-    fireEvent.click(screen.getByRole("button", { name: "환불 검증 replay" }));
-
-    expect(toast.error).toHaveBeenCalledWith("Replay version을 입력하세요.");
-    expect(mockedSimulationApi.replayGoldenCase).not.toHaveBeenCalled();
-  });
-
-  it("개선 후보 상태 필터를 변경해 목록을 다시 조회한다", async () => {
-    renderPage();
-
-    await openCandidateTab();
-    fireEvent.change(await screen.findByLabelText("개선 후보 상태 필터"), {
-      target: { value: "READY_FOR_REVIEW" },
-    });
-
-    await waitFor(() => {
-      expect(mockedSimulationApi.listImprovementCandidates).toHaveBeenCalledWith(1, {
-        status: "READY_FOR_REVIEW",
-        page: 0,
-        size: 20,
-      });
-    });
   });
 
   it("Enter 키로 고객 메시지를 전송한다", async () => {
